requirements.txt
cloudinary==1.44.1
mysql_connector_repackaged==0.3.1
pandas==2.3.3
plotly==5.24.1
streamlit==1.41.1
